усовершенствование загрузчика: загрузка одного списка категорий

+++ app/src/main/java/ru/neosvet/wallpaper/utils/GalleryLoader.java
@@ -26,6 +26,7 @@ import ru.neosvet.wallpaper.database.GalleryRepository;
 
 public class GalleryLoader extends IntentService implements LoaderMaster.IService {
     private final byte WORK = 0, SAVING = 1, FINISH = 2, ERROR = 3;
+    public static final int FINISH_ERROR = -1, FINISH_MINI = -2, FINISH_CATEGORIES = -3;
     private byte status = WORK;
     private final IBinder binder = new LoaderMaster.MyBinder(GalleryLoader.this);
     private int count = 0;
@@ -50,12 +51,11 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
 
     @Override
     protected void onHandleIntent(Intent intent) {
-        Lib.log("Thread: " + boolThread);
         if (site == null) {
             Settings settings = new Settings(GalleryLoader.this);
             site = settings.getSite();
         }
-        if (intent.hasExtra(DBHelper.URL)) {
+        if (intent.hasExtra(DBHelper.URL)) { // load mini
             status = WORK;
             urls.add(intent.getStringExtra(DBHelper.URL));
             if (repository == null)
@@ -83,18 +83,16 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
         String line;
         File f = new File(getFilesDir() + Lib.CATEGORIES);
         try {
+            //start:
             URL url = new URL(site + tag + "/page/" + page + "/");
-            Lib.log("url: "+url.toString());
             HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
             urlConnection.setConnectTimeout(Lib.cPing);
             urlConnection.setReadTimeout(Lib.cPing);
             //urlConnection.setRequestProperty(cUserAgent, cClient);
             InputStream in = new BufferedInputStream(urlConnection.getInputStream());
             BufferedReader br = new BufferedReader(new InputStreamReader(in, Lib.ENCODING), 1000);
-            repository = new GalleryRepository(GalleryLoader.this, DBHelper.LIST);
-            String url_image;
             line = br.readLine();
-
+            //categories:
             while (!line.contains("Sandbox"))
                 line = br.readLine();
             line = br.readLine();
@@ -113,7 +111,16 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
                 line = br.readLine();
             }
             bw.close();
-
+            if (page == 0) {
+                in.close();
+                urlConnection.disconnect();
+                status = FINISH;
+                count = FINISH_CATEGORIES;
+                return true;
+            }
+            //gallery:
+            String url_image;
+            repository = new GalleryRepository(GalleryLoader.this, DBHelper.LIST);
             while (!line.contains("col-md-4"))
                 line = br.readLine();
             while (!line.contains("<nav>") && !line.contains("<noindex>")) {
@@ -124,21 +131,13 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
                     url_image = line.substring(0, line.indexOf("\""));
                     repository.addUrl(url_image);
                     line = br.readLine();
-//                    if (line.contains("holder.js")) { //for tag
-//                        line = br.readLine();
-//                        urls.add(url_image);
-//                        if (!boolThread)
-//                            startThread();
-//                    } else {
-//                        line = line.substring(line.indexOf(site) + site.length() + 5, line.length() - 1);
-//                        repository.addMini(line);
-//                    }
                     if (line.contains("holder.js")) //for tag
                         line = br.readLine();
                     line = line.substring(line.indexOf(site) + site.length() + 5, line.length() - 1);
                     repository.addMini(line);
                 }
             }
+            //count pages:
             if (line.contains("<nav>")) {
                 line = br.readLine();
                 while (!line.contains("next") && !line.contains("last")) {
@@ -155,7 +154,7 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
                 }
                 count = Integer.parseInt(line);
             }
-
+            //finish:
             in.close();
             urlConnection.disconnect();
             status = SAVING;
@@ -164,8 +163,10 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
             if (urls.size() > 0 && !boolThread)
                 startThread();
         } catch (Exception e) {
-            f.delete();
+            if (f.exists())
+                f.delete();
             status = ERROR;
+            count = FINISH_ERROR;
             e.printStackTrace();
             return false;
         }
@@ -174,23 +175,17 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
 
     private void startThread() {
         boolThread = true;
-        Lib.log("startThread");
         new Thread(new Runnable() {
             @Override
             public void run() {
                 try {
                     String mini;
                     while (urls.size() > 0 && status != ERROR) {
-                        Lib.log("url: " + urls.get(0));
                         mini = getMini(urls.get(0));
-                        Lib.log("mini: " + mini);
-                        Lib.log("status: " + status);
                         while (status == SAVING) {
-                            Lib.log("sleep");
                             Thread.sleep(200);
                         }
                         if (status == FINISH) {
-                            Lib.log("status FINISH");
                             repository.updateMini(urls.get(0), mini);
                             if (act != null)
                                 act.updateGallery();
@@ -202,13 +197,12 @@ public class GalleryLoader extends IntentService implements LoaderMaster.IServic
                     e.printStackTrace();
                 }
                 boolThread = false;
-                Lib.log("finish Thread");
                 if (status == FINISH) {
                     if (act != null) {
                         act.runOnUiThread(new Runnable() {
                             @Override
                             public void run() {
-                                act.onPost(true, -1);
+                                act.onPost(true, FINISH_MINI);
                                 stopSelf();
                             }
                         });
